excel_to_mysql/requirements.txt
apache-airflow-providers-mysql>=5.5.0
pandas>=2.0.0
openpyxl>=3.1.0
SQLAlchemy>=1.4
PyMySQL>=1.1.0
pendulum>=2.1.2
